Maintenance commit: - Added some documentation, - Made modifications  to the build scripts to build the UX on Windows, Linux & MacOS. - Updated OTD.EnhancedOutputMode so that touch & this plugin can be used at the same time

+++ build.sh
@@ -14,9 +14,15 @@ else
     exit 1
 fi
 
-mkdir ./build/plugin
+if [ ! -d "./build" ]; then
+    mkdir ./build
+fi
 
-rm -rf ./build/plugin/*
+if [ ! -d "./build/plugin" ]; then
+    mkdir ./build/plugin
+else
+    rm -rf ./build/plugin/*
+fi
 
 mv ./temp/plugin/Wheel-Addon.dll ./build/plugin/Wheel-Addon.dll
 mv ./temp/plugin/Wheel-Addon.pdb ./build/plugin/Wheel-Addon.pdb
@@ -34,7 +40,17 @@ echo ""
 echo "Building UX"
 echo ""
 
-dotnet publish Wheel-Addon.UX.Desktop -c Release -r linux-x64 --self-contained="false" -o ./build/linux-x64/ $@ || exit 2
+rm -rf ./build/UX
+
+dotnet publish Wheel-Addon.UX.Desktop -c Release -r win-x64 --self-contained="false" -o ./build/UX/win-x64/ $@ || exit 2
+dotnet publish Wheel-Addon.UX.Desktop -c Release -r win-x86 --self-contained="false" -o ./build/UX/win-x86/ $@ || exit 2
+dotnet publish Wheel-Addon.UX.Desktop -c Release -r win-arm64 --self-contained="false" -o ./build/UX/win-arm64/ $@ || exit 2
+
+dotnet publish Wheel-Addon.UX.Desktop -c Release -r linux-x64 --self-contained="false" -o ./build/UX/linux-x64/ $@ || exit 2
+dotnet publish Wheel-Addon.UX.Desktop -c Release -r linux-arm64 --self-contained="false" -o ./build/UX/linux-arm64/ $@ || exit 2
+
+dotnet publish Wheel-Addon.UX.Desktop -c Release -r osx-x64 --self-contained="false" -o ./build/UX/osx-x64/ $@ || exit 2
+dotnet publish Wheel-Addon.UX.Desktop -c Release -r osx-arm64 --self-contained="false" -o ./build/UX/osx-arm64/ $@ || exit 2
 
 # if error is 0 then exit
 if [ $? -eq 0 ]; then
@@ -63,6 +79,10 @@ mkdir ./build/installer/
 rm -rf ./build/installer/*
 
 mv ./temp/installer/Wheel-Addon.Installer.dll ./build/installer/Wheel-Addon.Installer.dll
+
+# zip Wheel-Addon.Installer.dll
+
+zip -r ./build/installer/Wheel-Addon.Installer.zip ./build/installer/*
 
 rm -rf ./temp
